floder perpared

+++ client/src/components/Users/Index.vue
@@ -1,55 +1,76 @@
 <template>
   <div>
-    <h2>Get all users</h2>
-    <h4>จํานวนผู้ใช้งาน {{ users.length }}</h4>
-    <button v-on:click="navigateTo('/user/create')">สร้างผู้ใช้งาน</button>
-    <div v-for="user in users" v-bind:key="user.id">
-      <div><b>id: </b>{{ user.id }}</div>
-      <div><b>ชื่อผู้ใช้:</b> {{ user.name }} {{ user.lastname }}</div>
-      <div><b>email:</b> {{ user.email }}</div>
-      <div><b>status:</b> {{ user.status }}</div>
-      <div><b>type:</b> {{ user.type }}</div>
-      <div>
-        <button v-on:click="navigateTo('/user/' + user.id)">ดูข้อมูลผู้ใช้</button>
-        <button v-on:click="navigateTo('/user/edit/' + user.id)">แก้ไขข้อมูล</button>
-        <button v-on:click="deleteUser(user)">ลบข้อมูล</button>
+    <h1>All Users</h1>
+    <div><button v-on:click="navigateTo('/user/create')">สร้างผู้ใช้</button></div>
+    <hr>
+    <div v-if="users.length">
+      <div><b>จำนวนผู้ใช้งาน:</b> {{ users.length }}</div>
+      <div v-for="user in users" v-bind:key="user.id">
+        <div><b>id:</b> {{ user.id }}</div>
+        <div><b>ชื่อผู้ใช้:</b> {{ user.name }} {{ user.lastname }}</div>
+        <div><b>อีเมล:</b> {{ user.email }}</div>
+        <div><b>status:</b> {{ user.status }}</div>
+        <div><b>type:</b> {{ user.type }}</div>
+        <div>
+          <button v-on:click="navigateTo('/user/'+user.id)">ดูข้อมูล</button>
+          <button v-on:click="navigateTo('/user/edit/'+user.id)">แก้ไขข้อมูล</button>
+          <button v-on:click="deleteUser(user)">ลบข้อมูล</button>
+        </div>
+        <hr>
       </div>
-      <hr>
     </div>
+    <div><button v-on:click="logout">Logout</button></div>
+
   </div>
 </template>
-<script>
-import UsersService from '@/services/UsersService'
 
+<script>
+import UsersService from "@/services/UsersService";
 export default {
-  data() {
+  data(){
     return {
       users: []
     }
   },
   async created() {
-    this.users = (await UsersService.index()).data
+    try{
+      this.users = (await UsersService.index()).data;
+    }catch(err){
+      console.log(err);
+    }
   },
-  methods: {
-    navigateTo(route) {
-      this.$router.push(route)
+  methods:{
+    logout(){
+      this.$store.dispatch('setToken',null)
+      this.$store.dispatch('setUser',null)
+      this.$router.push({
+        name: 'login'
+      })
+    },
+    navigateTo(route){
+      this.$router.push(route);
     },
     async deleteUser(user){
-      let result = confirm("คุณต้องการลบข้อมูลใช่หรือไม่");
-      try{
-        await UsersService.delete(user);
-      } catch(err){
-        console.log(err);
+      let result = confirm("คุณต้องการลบข้อมูลใช่หรือไม่?");
+      if(result){
+        try{
+          await UsersService.delete(user);
+          this.refreshData();
+
+        }catch(err){
+          console.log(err);
+        }
       }
     },
     async refreshData(){
       try{
         this.users = (await UsersService.index()).data;
-      } catch(err){
-        console.log(err)
+      }catch(err){
+        console.log(err);
       }
     }
   }
-} 
+};
 </script>
-<style scoped></style>
+
+<style></style>
